refactor(cli): extract the shared adb executor
emulatorSetup and createAndroidEmulator each carried a copy of defaultAdb.
The comment in emulatorSetup asked for this once a third call site appeared.
