+++ src/shell/appium/emulatorSetup.ts
@@ -1,17 +1,4 @@
-import { execFile } from "node:child_process";
-import { promisify } from "node:util";
-
-import { adbBin } from "~/core/androidBins.js";
-import type { AdbFn } from "./createAndroidEmulator.js";
-
-const execFileAsync = promisify(execFile);
-
-// Duplicates the private defaultAdb in createAndroidEmulator.ts.
-export const defaultAdb: AdbFn = async (args) => {
-  const home = process.env["ANDROID_HOME"] ?? process.env["ANDROID_SDK_ROOT"];
-  const { stdout } = await execFileAsync(adbBin(home, process.platform), args);
-  return { stdout };
-};
+import type { AdbFn } from "./adb.js";
 
 async function disableAnimations(adb: AdbFn, serial: string): Promise<void> {
   await Promise.all([
